修复内容流处理中的错误信息格式，确保错误信息正确发送。更新全局样式中的主页面背景颜色，提升视觉一致性。重构ModernAnalysisInterface组件，使用PromptsContext管理prompts状态，简化代码并提升可维护性。移除不必要的prompts加载逻辑，优化用户体验。更新配置文件，调整模型名称以保持一致性。
