+++ litellm/config.yaml
@@ -18,19 +18,14 @@ general_settings:
 
 model_list:
   # OpenRouter Gemini Models
-  - model_name: or-gemini-2.5-flash-preview-05-20
-    litellm_params:
-      model: openrouter/google/gemini-2.5-flash-preview-05-20
-      api_key: "os.environ/OR_API_KEY"
-
   - model_name: or-gemini-2.5-pro
     litellm_params:
       model: openrouter/google/gemini-2.5-pro
       api_key: "os.environ/OR_API_KEY"
 
-  - model_name: or-gemini-2.0-flash
+  - model_name: or-gemini-2.5-flash
     litellm_params:
-      model: openrouter/google/gemini-2.0-flash
+      model: openrouter/google/gemini-2.5-flash
       api_key: "os.environ/OR_API_KEY"
 
   # OpenAI Models (暂时注释掉，因为连接有问题)
